requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.5
django-heroku==0.3.1
gunicorn==19.8.1
idna==2.7
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-decouple==3.1
pytz==2018.4
requests==2.19.1
urllib3==1.23
whitenoise==3.3.1
